+++ internal/dto/employee/update_request.go
@@ -6,7 +6,7 @@ type UpdateEmployeeRequest struct {
 	Password    *string `json:"password,omitempty" validate:"omitempty,min=8"`
 	Role        *string `json:"role,omitempty" validate:"omitempty,oneof=admin staff manager"`
 	Position    *string `json:"position,omitempty"`
-	Salary      *int    `json:"salary,omitempty" validate:"omitempty,gt=0"`
+	Salary      *int64  `json:"salary,omitempty" validate:"omitempty,gt=0"`
 	Status      *string `json:"status,omitempty" validate:"omitempty,oneof=active inactive"`
 	BirthDate   *string `json:"birth_date,omitempty" validate:"omitempty,datetime=2006-01-02"`
 	Address     *string `json:"address,omitempty"`
